requirements.txt
streamlit==1.31.0
pandas==2.2.0
python-dotenv==1.0.1
plotly 
wordcloud
matplotlib
pg8000
sqlalchemy
supabase
